prime factors

+++ 0x04-more_functions_nested_loops/100-prime_factor.c
@@ -12,7 +12,7 @@
 
 int main(void)
 {
-	int i, j, isPrime;
+	unsigned long i, j, isPrime;
 
 	for (i = 2; i <= 612852475143; i++)
 	{
